requirements.txt
psycopg2==2.8.5
pyyaml==5.1
validators==0.18.2
beautifulsoup4==4.9.3